move file where it's belongs to
